Divided code in pages
Le funzioni sono state divise nei vari file delle pagine. Nelle pagine di analisi c'è un bug. Non carica i databframe
