--- pages/01_➕ Home.py
@@ -1,8 +1,0 @@
-from functions import ciao
-
-import streamlit as st
-
-
-
-ciao()
-
